requirements.txt
openai
flask
flask-cors
python-dotenv
requests
gspread
google-auth
python-telegram-bot==20.7
gunicorn
